backend/app/requirements.txt
fastapi==0.115.8
langchain==0.3.19
langchain_core==0.3.37
langchain_openai==0.3.6
langgraph==0.2.74
marimo==0.11.8
pydantic==2.10.6
PyPDF2==3.0.1
python-dotenv==1.0.1
